[APM] Move index settings persistance to data access plugn
