+++ x-pack/plugins/apm/server/routes/settings/apm_indices/route.ts
@@ -8,12 +8,12 @@
 import * as t from 'io-ts';
 import { SavedObject } from '@kbn/core/server';
 import type { APMIndices } from '@kbn/apm-data-access-plugin/server';
+import { saveApmIndices } from '@kbn/apm-data-access-plugin/server/saved_objects/apm_indices';
 import { createApmServerRoute } from '../../apm_routes/create_apm_server_route';
 import {
   getApmIndexSettings,
   ApmIndexSettingsResponse,
 } from './get_apm_indices';
-import { saveApmIndices } from './save_apm_indices';
 
 // get list of apm indices and values
 const apmIndexSettingsRoute = createApmServerRoute({
