+++ x-pack/plugins/apm_data_access/server/saved_objects/apm_indices.ts
@@ -11,6 +11,7 @@ import { schema } from '@kbn/config-schema';
 import { SavedObjectsErrorHelpers } from '@kbn/core/server';
 import { SavedObjectsClientContract } from '@kbn/core/server';
 import { updateApmOssIndexPaths } from './migrations/update_apm_oss_index_paths';
+import { APMIndices } from '..';
 
 export const APM_INDEX_SETTINGS_SAVED_OBJECT_TYPE = 'apm-indices';
 export const APM_INDEX_SETTINGS_SAVED_OBJECT_ID = 'apm-indices';
@@ -72,6 +73,28 @@ export const apmIndicesSavedObjectDefinition: SavedObjectsType = {
     },
   },
 };
+
+export function saveApmIndices(
+  savedObjectsClient: SavedObjectsClientContract,
+  apmIndices: Partial<APMIndices>
+) {
+  return savedObjectsClient.create<APMIndicesSavedObjectBody>(
+    APM_INDEX_SETTINGS_SAVED_OBJECT_TYPE,
+    { apmIndices: removeEmpty(apmIndices), isSpaceAware: true },
+    { id: APM_INDEX_SETTINGS_SAVED_OBJECT_ID, overwrite: true }
+  );
+}
+
+// remove empty/undefined values
+function removeEmpty(apmIndices: Partial<APMIndices>) {
+  return Object.entries(apmIndices)
+    .map(([key, value]) => [key, value?.trim()])
+    .filter(([_, value]) => !!value)
+    .reduce((obj, [key, value]) => {
+      obj[key] = value;
+      return obj;
+    }, {} as Record<string, unknown>);
+}
 
 export async function getApmIndicesSavedObject(savedObjectsClient: SavedObjectsClientContract) {
   try {
